backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
anthropic>=0.25.0
requests>=2.28.0
twilio>=8.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0
supabase>=2.0.0
websockets>=12.0,<16.0
docusign-esign>=3.24.0
fpdf2>=2.7.9
Pillow>=10.0.0
pydantic[email]>=2.0.0
APScheduler>=3.10.0
resend>=0.7.0
